+++ src/main/java/com/atlauncher/Gsons.java
@@ -21,6 +21,7 @@ import java.awt.Color;
 import java.time.Instant;
 import java.util.Date;
 
+import com.atlauncher.annot.ExcludeFromGsonSerialization;
 import com.atlauncher.data.AbstractAccount;
 import com.atlauncher.data.AccountTypeAdapter;
 import com.atlauncher.data.ColorTypeAdapter;
@@ -40,22 +41,38 @@ import com.atlauncher.data.minecraft.loaders.forge.ForgeLibrary;
 import com.atlauncher.data.minecraft.loaders.forge.ForgeLibraryTypeAdapter;
 import com.atlauncher.data.minecraft.loaders.quilt.QuiltMetaLauncherMeta;
 import com.atlauncher.data.minecraft.loaders.quilt.QuiltMetaLauncherMetaTypeAdapter;
+import com.google.gson.ExclusionStrategy;
+import com.google.gson.FieldAttributes;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 
 public final class Gsons {
+    public static final ExclusionStrategy exclusionAnnotationStrategy = new ExclusionStrategy() {
+        @Override
+        public boolean shouldSkipClass(Class<?> clazz) {
+            return false;
+        }
+
+        @Override
+        public boolean shouldSkipField(FieldAttributes field) {
+            return field.getAnnotation(ExcludeFromGsonSerialization.class) != null;
+        }
+    };
+
     private static final Gson BASE = new GsonBuilder()
             .registerTypeAdapter(AbstractAccount.class, new AccountTypeAdapter())
             .registerTypeAdapter(Date.class, new DateTypeAdapter())
             .registerTypeAdapter(Color.class, new ColorTypeAdapter())
-            .registerTypeAdapter(OauthTokenResponse.class, new OauthTokenResponseTypeAdapter()).create();
+            .registerTypeAdapter(OauthTokenResponse.class, new OauthTokenResponseTypeAdapter())
+            .addSerializationExclusionStrategy(exclusionAnnotationStrategy).create();
 
     public static final Gson DEFAULT = BASE.newBuilder().setPrettyPrinting().create();
 
     public static final Gson DEFAULT_SLIM = BASE.newBuilder().create();
 
     public static final Gson DEFAULT_ALT = new GsonBuilder().registerTypeAdapter(Color.class, new ColorTypeAdapter())
-            .registerTypeAdapter(PackVersion.class, new PackVersionTypeAdapter()).setPrettyPrinting().create();
+            .registerTypeAdapter(PackVersion.class, new PackVersionTypeAdapter())
+            .addSerializationExclusionStrategy(exclusionAnnotationStrategy).setPrettyPrinting().create();
 
     public static final Gson MINECRAFT = new GsonBuilder().disableHtmlEscaping().setPrettyPrinting()
             .registerTypeAdapter(Instant.class, new InstantTypeAdapter())
@@ -64,5 +81,6 @@ public final class Gsons {
             .registerTypeAdapter(Arguments.class, new ArgumentsTypeAdapter())
             .registerTypeAdapter(FabricMetaLauncherMeta.class, new FabricMetaLauncherMetaTypeAdapter())
             .registerTypeAdapter(ForgeLibrary.class, new ForgeLibraryTypeAdapter())
-            .registerTypeAdapter(QuiltMetaLauncherMeta.class, new QuiltMetaLauncherMetaTypeAdapter()).create();
+            .registerTypeAdapter(QuiltMetaLauncherMeta.class, new QuiltMetaLauncherMetaTypeAdapter())
+            .addSerializationExclusionStrategy(exclusionAnnotationStrategy).create();
 }