+++ src/main/java/com/atlauncher/data/DisableableMod.java
@@ -324,9 +324,13 @@ public class DisableableMod implements Serializable {
     }
 
     public boolean checkForUpdate(Window parent, Instance instance) {
+        return checkForUpdate(parent, instance, null);
+    }
+
+    public boolean checkForUpdate(Window parent, Instance instance, ModPlatform platform) {
         Analytics.sendEvent(instance.launcher.pack + " - " + instance.launcher.version, "UpdateMods", "Instance");
 
-        if (isFromCurseForge()) {
+        if (platform == ModPlatform.CURSEFORGE || platform == null && isFromCurseForge()) {
             ProgressDialog<Object> dialog = new ProgressDialog<>(
                     GetText.tr("Checking For Update On CurseForge"), 0, GetText.tr("Checking For Update On CurseForge"),
                     "Cancelled checking for update on CurseForge", parent);
@@ -392,7 +396,7 @@ public class DisableableMod implements Serializable {
 
             new CurseForgeProjectFileSelectorDialog(parent, (CurseForgeProject) dialog.getReturnValue(), instance,
                     curseForgeFileId);
-        } else if (isFromModrinth()) {
+        } else if (platform == ModPlatform.MODRINTH || platform == null && isFromModrinth()) {
             ProgressDialog<Pair<ModrinthProject, List<ModrinthVersion>>> dialog = new ProgressDialog<>(
                     GetText.tr("Checking For Update On Modrinth"), 0,
                     GetText.tr("Checking For Update On Modrinth"), "Cancelled checking for update on Modrinth", parent);
@@ -439,9 +443,13 @@ public class DisableableMod implements Serializable {
     }
 
     public boolean reinstall(Window parent, Instance instance) {
+        return reinstall(parent, instance, null);
+    }
+
+    public boolean reinstall(Window parent, Instance instance, ModPlatform platform) {
         Analytics.sendEvent(instance.launcher.pack + " - " + instance.launcher.version, "ReinstallMods", "Instance");
 
-        if (isFromCurseForge()) {
+        if (platform == ModPlatform.CURSEFORGE || platform == null && isFromCurseForge()) {
             ProgressDialog<CurseForgeProject> dialog = new ProgressDialog<>(GetText.tr("Getting Files From CurseForge"),
                     0, GetText.tr("Getting Files From CurseForge"), "Cancelled getting files from CurseForge", parent);
             dialog.addThread(new Thread(() -> {
@@ -451,7 +459,7 @@ public class DisableableMod implements Serializable {
             dialog.start();
 
             new CurseForgeProjectFileSelectorDialog(parent, dialog.getReturnValue(), instance, curseForgeFileId);
-        } else if (isFromModrinth()) {
+        } else if (platform == ModPlatform.MODRINTH || platform == null && isFromModrinth()) {
             ProgressDialog<ModrinthProject> dialog = new ProgressDialog<>(GetText.tr("Getting Files From Modrinth"), 0,
                     GetText.tr("Getting Files From Modrinth"), "Cancelled getting files from Modrinth", parent);
             dialog.addThread(new Thread(() -> {