+++ src/main/java/com/atlauncher/gui/components/ModsJCheckBox.java
@@ -27,6 +27,7 @@ import javax.swing.JToolTip;
 
 import com.atlauncher.builders.HTMLBuilder;
 import com.atlauncher.data.DisableableMod;
+import com.atlauncher.data.ModPlatform;
 import com.atlauncher.data.json.Mod;
 import com.atlauncher.data.modrinth.ModrinthDonationUrl;
 import com.atlauncher.data.modrinth.ModrinthProject;
@@ -226,7 +227,57 @@ public class ModsJCheckBox extends JCheckBox {
         });
         contextMenu.add(remove);
 
-        if (getDisableableMod().isFromCurseForge() || getDisableableMod().isFromModrinth()) {
+        if (getDisableableMod().isFromCurseForge() && getDisableableMod().isFromModrinth()) {
+            contextMenu.add(new JPopupMenu.Separator());
+
+            JMenuItem reinstallFromCurseForge = new JMenuItem(GetText.tr("Reinstall From CurseForge"));
+            reinstallFromCurseForge.addActionListener(e -> {
+                getDisableableMod().reinstall(dialog, dialog.instance, ModPlatform.CURSEFORGE);
+
+                dialog.reloadPanels();
+            });
+            contextMenu.add(reinstallFromCurseForge);
+
+            JMenuItem reinstallFromModrinth = new JMenuItem(GetText.tr("Reinstall From Modrinth"));
+            reinstallFromModrinth.addActionListener(e -> {
+                getDisableableMod().reinstall(dialog, dialog.instance, ModPlatform.MODRINTH);
+
+                dialog.reloadPanels();
+            });
+            contextMenu.add(reinstallFromModrinth);
+
+            contextMenu.add(new JPopupMenu.Separator());
+
+            JMenuItem checkForUpdatesOnCurseForge = new JMenuItem(GetText.tr("Check For Updates On CurseForge"));
+            checkForUpdatesOnCurseForge.addActionListener(e -> {
+                boolean updated = false;
+
+                updated = getDisableableMod().checkForUpdate(dialog, dialog.instance, ModPlatform.CURSEFORGE);
+
+                if (!updated) {
+                    DialogManager.okDialog().setTitle(GetText.tr("No Updates Found"))
+                            .setContent(GetText.tr("No updates were found.")).show();
+                }
+
+                dialog.reloadPanels();
+            });
+            contextMenu.add(checkForUpdatesOnCurseForge);
+
+            JMenuItem checkForUpdatesOnModrinth = new JMenuItem(GetText.tr("Check For Updates On Modrinth"));
+            checkForUpdatesOnModrinth.addActionListener(e -> {
+                boolean updated = false;
+
+                updated = getDisableableMod().checkForUpdate(dialog, dialog.instance, ModPlatform.MODRINTH);
+
+                if (!updated) {
+                    DialogManager.okDialog().setTitle(GetText.tr("No Updates Found"))
+                            .setContent(GetText.tr("No updates were found.")).show();
+                }
+
+                dialog.reloadPanels();
+            });
+            contextMenu.add(checkForUpdatesOnModrinth);
+        } else if (getDisableableMod().isFromCurseForge() || getDisableableMod().isFromModrinth()) {
             contextMenu.add(new JPopupMenu.Separator());
 
             JMenuItem reinstall = new JMenuItem(GetText.tr("Reinstall"));