+++ CHANGELOG.md
@@ -11,5 +11,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Remove caching of download urls from CurseForge as per ToS, also remove some others from json
 - Don't export CurseForge mods urls in Modrinth pack exports due to CurseForge ToS, use overides
 - When exporting to Modrinth format, check all mods on Modrinth
+- Change right click menu on edit mods for mods with info from both mod platforms
 
 ### Misc