+++ CHANGELOG.md
@@ -3,12 +3,10 @@
 This changelog only contains the changes that are unreleased. For changes for individual releases, please visit the
 [releases](https://github.com/ATLauncher/ATLauncher/releases) page on GitHub.
 
-## 3.4.13.8
+## 3.4.13.9
 
 ### New Features
 
 ### Fixes
-- Update Modrinth icon and support dark and light modpack platform icons
-- Add missing fileSize from Modrinth modpack spec
 
 ### Misc